Fix for issue 295 reworking Fortran and C bindings
Reorganized bindings components to map ADIOS2 core classes
Using cmake FortranCInterface to automatically resolve name mangling
