+++ bindings/C/CMakeLists.txt
@@ -4,8 +4,13 @@
 #------------------------------------------------------------------------------#
 
 
-target_sources(adios2 PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/adios2_c.cpp
-${CMAKE_CURRENT_SOURCE_DIR}/adios2/adios2_c_glue.cpp)
+target_sources(adios2 PRIVATE 
+${CMAKE_CURRENT_SOURCE_DIR}/adios2/adios2_c_adios.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/adios2/adios2_c_glue.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/adios2/adios2_c_io.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/adios2/adios2_c_engine.cpp
+)
+
 target_include_directories(adios2
   PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
@@ -15,7 +20,12 @@ install(
   FILES adios2_c.h 
   DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
+
 install(
-  FILES adios2/adios2_c_types.h adios2/adios2_c_glue.h
+  FILES adios2/adios2_c_types.h 
+        adios2/adios2_c_adios.h
+        adios2/adios2_c_glue.h 
+        adios2/adios2_c_io.h
+        adios2/adios2_c_engine.h
   DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}/adios2
 )
