+++ bindings/fortran/adios2_mod.f90
@@ -10,10 +10,10 @@
 
 module adios2
 
-    use adios2_params
+    use adios2_parameters
     use adios2_adios
     use adios2_io
-!    use adios2_variable
+!    TODO use adios2_variable
     use adios2_engine
 
 end module
