+++ bindings/fortran/modules/nompi/adios2_io_open_mod.f90
@@ -2,7 +2,7 @@
 ! Distributed under the OSI-approved Apache License, Version 2.0.  See
 !  accompanying file Copyright.txt for details.
 !
-!  adios2_io_open_nompi_mod.f90 : ADIOS2 Fortran bindings for IO class open
+!  adios2_io_open_mod.f90 : ADIOS2 Fortran bindings for IO class open
 !                                 function
 !   Created on: Mar 13, 2017
 !       Author: William F Godoy godoywf@ornl.gov
